backend/requirements/base.txt
# Dependencies used in: Development, Production, Testing
asgiref==3.9.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
cryptography==46.0.2
Django==5.2.7

# Authentication / Registration
dj-rest-auth[with_social]==7.0.1
django-allauth==65.13.0

django-cors-headers==4.3.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
gunicorn==21.2.0
idna==3.10
packaging==25.0
psycopg2-binary==2.9.9
pycparser==2.23
PyJWT==2.10.1
python-dotenv==1.0.0
requests==2.32.5
sqlparse==0.5.3
urllib3==2.5.0
python-decouple==3.8

# Real-time via Django Channels
channels==4.1.0
channels-redis==4.2.0

# ASGI server for Django/Channels
daphne==4.2.1
uvicorn[standard]

# PDF generation
xhtml2pdf==0.2.8
reportlab==3.6.13

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0